+++ RASPCompiler/src/sk/tuke/emustudio/rasp/compiler/tree/Statement.java
@@ -5,7 +5,7 @@
  */
 package sk.tuke.emustudio.rasp.compiler.tree;
 
-import sk.tuke.emustudio.rasp.memory.impl.RASPInstructionImpl;
+import sk.tuke.emustudio.rasp.memory.memoryitems.RASPInstructionImpl;
 
 /**
  *
