+++ RASPMemory/src/sk/tuke/emustudio/rasp/memory/memoryitems/NumberMemoryItem.java
@@ -1,4 +1,4 @@
-package sk.tuke.emustudio.rasp.memory;
+package sk.tuke.emustudio.rasp.memory.memoryitems;
 
 /**
  * Value as a memory item, either value of register, or instruction operand.
