+++ RASPMemory/src/sk/tuke/emustudio/rasp/memory/memoryitems/RASPInstruction.java
@@ -1,4 +1,4 @@
-package sk.tuke.emustudio.rasp.memory;
+package sk.tuke.emustudio.rasp.memory.memoryitems;
 
 /**
  * Interface representing RASP instruction.
